requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.0+cu117
torchvision==0.15.1+cu117
torchaudio==2.0.1
hydra-core==1.0.5
timm
diffusers
accelerate
omegaconf
wandb
matplotlib
python-dotenv
scipy